# perf(css): Optimize memory layout of ast types