--- a/crates/swc_css_ast/src/at_rule.rs
+++ b/crates/swc_css_ast/src/at_rule.rs
@@ -309,7 +309,7 @@ pub enum MediaInParens {
     MediaCondition(MediaCondition),
 
     #[tag("MediaFeature")]
-    Feature(MediaFeature),
+    Feature(Box<MediaFeature>),
     // TODO <general-enclosed>
 }