--- a/crates/swc_css_lints/src/lib.rs
+++ b/crates/swc_css_lints/src/lib.rs
@@ -1,4 +1,5 @@
 #![deny(clippy::all)]
+#![feature(box_patterns)]
 
 mod config;
 mod dataset;