--- a/crates/swc_css_lints/src/rules/at_rule_no_unknown.rs
+++ b/crates/swc_css_lints/src/rules/at_rule_no_unknown.rs
@@ -40,7 +40,7 @@ struct AtRuleNoUnknown {
 impl Visit for AtRuleNoUnknown {
     fn visit_at_rule(&mut self, at_rule: &AtRule) {
         if let AtRuleName::Ident(Ident { value, .. }) = &at_rule.name {
-            if let Some(AtRulePrelude::ListOfComponentValues(_)) = at_rule.prelude {
+            if let Some(AtRulePrelude::ListOfComponentValues(_)) = at_rule.prelude.as_deref() {
                 if self.ignored.iter().all(|item| !item.is_match(&value)) {
                     let message = format!("Unexpected unknown at-rule \"@{}\".", &value);