--- a/crates/swc_css_parser/src/parser/at_rule.rs
+++ b/crates/swc_css_parser/src/parser/at_rule.rs
@@ -322,7 +322,7 @@ where
                     let prelude = AtRulePrelude::NamespacePrelude(NamespacePrelude {
                         span: span!(parser, span.lo),
                         prefix,
-                        uri,
+                        uri: Box::new(uri),
                     });
 
                     if !is!(parser, ";") {