--- a/crates/swc_css_prefixer/src/lib.rs
+++ b/crates/swc_css_prefixer/src/lib.rs
@@ -1,3 +1,6 @@
+#![feature(box_patterns)]
+#![allow(clippy::vec_box)]
+
 pub use self::prefixer::prefixer;
 
 pub mod options;