**Description:**

```
Benchmarking css/minify/libraries/bootstrap
Benchmarking css/minify/libraries/bootstrap: Warming up for 3.0000 s
Benchmarking css/minify/libraries/bootstrap: Collecting 10 samples in estimated 5.2997 s (825 iterations)
Benchmarking css/minify/libraries/bootstrap: Analyzing
css/minify/libraries/bootstrap
                        time:   [6.4418 ms 6.4475 ms 6.4558 ms]
                        change: [-21.738% -21.403% -21.047%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
```